Move app tests to module package to allow for better test coverage
