+++ pkg/services/app/app_suite_test.go
@@ -1,4 +1,4 @@
-package app_test
+package app
 
 import (
 	"testing"
