+++ pkg/services/app/get_test.go
@@ -1,4 +1,4 @@
-package app_test
+package app
 
 import (
 	"context"
